pyproject.toml
[project]
name = "todo-fast"
version = "0.1.0"
description = "Simple todo app with FastAPI and Streamlit."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi[standard]>=0.115.6",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.0",
]
